+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- CLI command to emulate stop button being pressed.
 
 
 ## [0.6.1] - 2022-04-27